+++ jest/mocks/dataMocks/channels/get.js
@@ -1,0 +1,3 @@
+import { getRoom } from '../rooms/getRooms'
+
+export const getChannel = () => getRoom()
