--- jest/mocks/dataMocks/get.js
@@ -1,3 +1,0 @@
-import { getRoom } from './getRooms'
-
-export const getChannel = () => getRoom()
