+++ jest/mocks/dataMocks/index.js
@@ -1,14 +1,14 @@
-import { getUsers } from './getUsers'
-import { getRooms } from './getRooms'
-import { create } from './create'
-import { getChannel } from './get'
+import { getUsers } from './users/getUsers'
+import { getRooms } from './rooms/getRooms'
+import { create } from './rooms/create'
+import { getChannel } from './channels/get'
 
-export { getUser, getUsers } from './getUsers'
-export { getRoom, getRooms } from './getRooms'
+export { getUser, getUsers } from './users/getUsers'
+export { getRoom, getRooms } from './rooms/getRooms'
 
 export default {
-  getUsers,
-  getRooms,
-  get: getChannel,
-  create,
+  usersGetUsers: getUsers,
+  roomsGetRooms: getRooms,
+  roomsCreate: create,
+  channelsGet: getChannel,
 }
